refactor: port appendToPlugin to typescript
