+++ packages/react/src/components/DatePicker/DatePicker.tsx
@@ -20,7 +20,7 @@ import cx from 'classnames';
 import flatpickr from 'flatpickr';
 import l10n from 'flatpickr/dist/l10n/index';
 import DatePickerInput from '../DatePickerInput';
-import carbonFlatpickrAppendToPlugin from './plugins/appendToPlugin';
+import { appendToPlugin } from './plugins/appendToPlugin';
 import carbonFlatpickrFixEventsPlugin from './plugins/fixEventsPlugin';
 import carbonFlatpickrRangePlugin from './plugins/rangePlugin';
 import deprecate from '../../prop-types/deprecate';
@@ -212,7 +212,7 @@ export interface DatePickerProps {
   /**
    * The DOM element the flatpickr should be inserted into `<body>` by default.
    */
-  appendTo?: object;
+  appendTo?: HTMLElement;
 
   /**
    * The child nodes.
@@ -645,7 +645,7 @@ const DatePicker = React.forwardRef(function DatePicker(
             })
           : () => {},
         appendTo
-          ? carbonFlatpickrAppendToPlugin({
+          ? appendToPlugin({
               appendTo,
             })
           : () => {},
